lvalue-width: CFaderRadialCell::m_pixel is u8; ImageRotateBlit's quad is wound
Two byte-evidenced correctness fixes plus the negative result that kills a
whole family of mask-sieve rows.

CFaderRadialCell::m_pixel was modelled i32.  Retail reads it with
`mov al,BYTE PTR [ebx+0xc]` (CFaderRadial::RenderFrame 0x17fd2a) and writes it
with a byte store (ApplyInit 0x17fbc9), then copies the cell out with four
DWORD moves that carry the three pad bytes - so the field is u8 and the write
is a whole-struct assignment from a local temp, not four member stores.  A
probe matrix (u8-vs-i32 field x struct-copy-vs-field-store) reproduces retail
only for the u8 + struct-copy cell.  ApplyInit also computes m_vy from
`m_centerY - y`, not `-dy`: retail spends a separate `sub ecx,ebp` for it.
86.39 -> 91.60.

ImageRotateBlit built its quad in RASTER order (TL, TR, BL, BR) - a bowtie -
with u taken from sq[3] and v from sq[2].  Retail's mtx store order (the
0x146097 loop) and its u/v tail (0x1460e7) pin the quad as TL, TR, BR, BL with
u from sq[0]/sq[2] and v from sq[1]/sq[3].  It also multiplies by the NEGATIVE
degrees-to-radians constant at 0x1efb14, which was defined-but-unreferenced in
the tree (so its DATA pin never bound); the source used a positive literal, so
the rotation ran the wrong way.  Renamed to g_degToRadNeg and moved into its
retail .rdata slot, which also puts the five pooled floats in retail's order.
36.42 -> 79.47.

`andl $0xff` on a stack aggregate's byte member is NOT a wrong-lvalue bug.
Six probe variants (union/plain field, memcpy/struct-assign/copy-init, pointer
and reference to the local, inlined ref and ptr helpers, array decay) all show
cl 5.0 loading a DWORD and masking for any byte member of a FRAME-LOCAL
aggregate, and `xor r,r; mov r8,m8` only for a genuine pointer deref.
CBattlezMapConfig::StepRowSpawn's `m_occupantIdBytes[1]` is byte-correct.

+++ include/Gruntz/FaderSubtypes.h
@@ -17,11 +17,15 @@ GZ_ENUM_FORWARD(FaderMode);
 class CDDSurface;
 struct CDDPalette;
 
+// m_pixel is a BYTE, not an i32: CFaderRadial::RenderFrame reads it with
+// `mov al,BYTE PTR [ebx+0xc]` (0x17fd2a) and ApplyInit writes it with a byte
+// store (0x17fbc9).  The three trailing pad bytes are why ApplyInit's cell
+// write is a 4-dword struct copy - it carries the padding too.
 struct CFaderRadialCell {
     float m_vx;
     float m_vy;
     float m_radius;
-    i32 m_pixel;
+    u8 m_pixel;
 };
 SIZE(0x10);
